moved code from test class to semantic class
