+++ semantic-convention-utils/src/main/java/org/hypertrace/semantic/convention/utils/rpc/RpcSemanticConventionUtils.java
@@ -1,5 +1,12 @@
 package org.hypertrace.semantic.convention.utils.rpc;
 
+import static org.hypertrace.core.span.constants.v1.CensusResponse.CENSUS_RESPONSE_STATUS_MESSAGE;
+import static org.hypertrace.core.span.constants.v1.Envoy.ENVOY_GRPC_STATUS_MESSAGE;
+import static org.hypertrace.core.span.constants.v1.Grpc.*;
+import static org.hypertrace.core.span.constants.v1.Grpc.GRPC_RESPONSE_BODY;
+import static org.hypertrace.core.span.normalizer.constants.RpcSpanTag.RPC_REQUEST_METADATA_AUTHORITY;
+import static org.hypertrace.core.span.normalizer.constants.RpcSpanTag.RPC_REQUEST_METADATA_USER_AGENT;
+
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 import java.util.List;
@@ -44,6 +51,11 @@ public class RpcSemanticConventionUtils {
           RawSpanConstants.getValue(CensusResponse.CENSUS_RESPONSE_CENSUS_STATUS_CODE));
   private static final String OTEL_RPC_SERVICE =
       OTelRpcSemanticConventions.RPC_SYSTEM_SERVICE.getValue();
+
+  private static final List<String> Status_Msg_Att =
+      List.of(
+          RawSpanConstants.getValue(CENSUS_RESPONSE_STATUS_MESSAGE),
+          RawSpanConstants.getValue(ENVOY_GRPC_STATUS_MESSAGE));
 
   /** @return attribute keys for grpc method */
   public static List<String> getAttributeKeysForGrpcMethod() {
@@ -115,6 +127,90 @@ public class RpcSemanticConventionUtils {
       return Optional.of(valueMap.get(OTHER_GRPC_HOST_PORT).getValue());
     } else if (isRpcTypeGrpcForOTelFormat(valueMap)) {
       return SpanSemanticConventionUtils.getURIForOtelFormat(valueMap);
+    }
+    return Optional.empty();
+  }
+
+  public static int getGrpcStatusCode(Event event) {
+    Map<String, AttributeValue> attributeValueMap = event.getAttributes().getAttributeMap();
+    for (String statuscode : RpcSemanticConventionUtils.getAttributeKeysForGrpcStatusCode()) {
+      if (attributeValueMap.get(statuscode) != null) {
+        return Integer.parseInt(attributeValueMap.get(statuscode).getValue());
+      }
+    }
+    return -1;
+  }
+
+  public static String getGrpcStatusMsg(Event event) {
+    Map<String, AttributeValue> attributeValueMap = event.getAttributes().getAttributeMap();
+    for (String statusmsg : Status_Msg_Att) {
+      if (attributeValueMap.get(statusmsg) != null) {
+        return attributeValueMap.get(statusmsg).getValue();
+      }
+    }
+    return "";
+  }
+
+  public static String getGrpcErrorMsg(Event event) {
+    Map<String, AttributeValue> attributeValueMap = event.getAttributes().getAttributeMap();
+    if (attributeValueMap.get(RawSpanConstants.getValue(GRPC_ERROR_MESSAGE)) != null) {
+      return attributeValueMap.get(RawSpanConstants.getValue(GRPC_ERROR_MESSAGE)).getValue();
+    }
+    return "";
+  }
+
+  public static Optional<String> getGrpcUserAgent(Event event) {
+    if (event.getAttributes() == null || event.getAttributes().getAttributeMap() == null) {
+      return Optional.empty();
+    }
+    ;
+    Map<String, AttributeValue> attributeValueMap = event.getAttributes().getAttributeMap();
+    if (attributeValueMap.get(RPC_REQUEST_METADATA_USER_AGENT.getValue()) != null
+        && !io.micrometer.core.instrument.util.StringUtils.isEmpty(
+            attributeValueMap.get(RPC_REQUEST_METADATA_USER_AGENT.getValue()).getValue())) {
+      return Optional.of(
+          attributeValueMap.get(RPC_REQUEST_METADATA_USER_AGENT.getValue()).getValue());
+    }
+    return Optional.empty();
+  }
+
+  public static Optional<String> getGrpcAuthority(Event event) {
+    if (event.getAttributes() == null || event.getAttributes().getAttributeMap() == null) {
+      return Optional.empty();
+    }
+    ;
+    Map<String, AttributeValue> attributeValueMap = event.getAttributes().getAttributeMap();
+    if (attributeValueMap.get(RPC_REQUEST_METADATA_AUTHORITY.getValue()) != null) {
+      return Optional.of(
+          attributeValueMap.get(RPC_REQUEST_METADATA_AUTHORITY.getValue()).getValue());
+    }
+    return Optional.empty();
+  }
+
+  public static Optional<Integer> getGrpcRequestSize(Event event) {
+    if (event.getAttributes() == null || event.getAttributes().getAttributeMap() == null) {
+      return Optional.empty();
+    }
+    ;
+    Map<String, AttributeValue> attributeValueMap = event.getAttributes().getAttributeMap();
+    if (attributeValueMap.get(RawSpanConstants.getValue(GRPC_REQUEST_BODY)) != null) {
+      return Optional.of(
+          Integer.parseInt(
+              attributeValueMap.get(RawSpanConstants.getValue(GRPC_REQUEST_BODY)).getValue()));
+    }
+    return Optional.empty();
+  }
+
+  public static Optional<Integer> getGrpcResponseSize(Event event) {
+    if (event.getAttributes() == null || event.getAttributes().getAttributeMap() == null) {
+      return Optional.empty();
+    }
+    ;
+    Map<String, AttributeValue> attributeValueMap = event.getAttributes().getAttributeMap();
+    if (attributeValueMap.get(RawSpanConstants.getValue(GRPC_RESPONSE_BODY)) != null) {
+      return Optional.of(
+          Integer.parseInt(
+              attributeValueMap.get(RawSpanConstants.getValue(GRPC_RESPONSE_BODY)).getValue()));
     }
     return Optional.empty();
   }
